Siddhesh_Badnapurkar_RBT23CB016/requirements.txt
ipykernel>=7.1.0
matplotlib>=3.10.7
numpy>=2.3.3
nvidia-cublas>=13.0.0.19
nvidia-cuda-cupti>=13.0.48
nvidia-cuda-nvrtc>=13.0.48
nvidia-cuda-runtime>=13.0.48
nvidia-cudnn-cu13>=9.13.0.50
nvidia-cufft>=12.0.0.15
nvidia-cufile>=1.15.0.42
nvidia-curand>=10.4.0.35
nvidia-cusolver>=12.0.3.29
nvidia-cusparse>=12.6.2.49
nvidia-cusparselt-cu13>=0.8.0
nvidia-nccl-cu13>=2.27.7
nvidia-nvjitlink>=13.0.39
nvidia-nvshmem-cu13>=3.3.24
nvidia-nvtx>=13.0.39
pillow>=11.3.0
sympy>=1.14.0
torch>=2.9.0+cu130
torchvision>=0.24.0+cu130
